// frontend/src/api-generated/models/donViResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface DonViResponse {
  id?: number;
  maDonVi?: string;
  tenPhapLy?: string;
  tenThuongMai?: string;
  maSoThue?: string;
  maQuocGiaDienThoai?: string;
  soDienThoaiCoDinh?: string;
  soDienThoaiDiDong?: string;
  emailChinhThuc?: string;
  tenMienHeThong?: string;
  duongDanWebsite?: string;
  soNhaTenDuong?: string;
  phuongXa?: string;
  quanHuyen?: string;
  tinhThanhPho?: string;
  maBuuChinh?: string;
  maQuocGia?: string;
  hoNguoiDaiDien?: string;
  tenNguoiDaiDien?: string;
  tenDemNguoiDaiDien?: string;
  chucVuNguoiDaiDien?: string;
  trangThai?: string;
  thoiGianThanhLap?: string;
  thoiGianBatDauHopDong?: string;
  thoiGianHetHanHopDong?: string;
  thoiGianTao?: string;
  thoiGianCapNhat?: string;
}